expose modifier utility function to consumer
