+++ test/Helpers.js
@@ -57,214 +57,6 @@ describe('Helpers', () => {
     });
   });
 
-  describe('getModifiersForDay', () => {
-    it('returns an array of modifiers', () => {
-      const modifierFunctions = {
-        yes() {
-          return true;
-        },
-        no() {
-          return false;
-        },
-        maybe(d) {
-          return d.getMonth() === 8;
-        },
-      };
-      let modifiers = Helpers.getModifiersForDay(
-        new Date(2015, 8, 19),
-        modifierFunctions
-      );
-      expect(modifiers).to.have.length(2);
-      expect(modifiers.indexOf('yes')).to.be.above(-1);
-      expect(modifiers.indexOf('maybe')).to.be.above(-1);
-      expect(modifiers.indexOf('no')).to.equal(-1);
-
-      modifiers = Helpers.getModifiersForDay(
-        new Date(2015, 9, 19),
-        modifierFunctions
-      );
-      expect(modifiers).to.have.length(1);
-      expect(modifiers.indexOf('yes')).to.be.above(-1);
-      expect(modifiers.indexOf('maybe')).to.equal(-1);
-      expect(modifiers.indexOf('no')).to.equal(-1);
-    });
-    it('returns the modifier for a single day', () => {
-      const modifiers = Helpers.getModifiersForDay(new Date(2015, 8, 19), {
-        foo: new Date(2015, 8, 19),
-      });
-      expect(modifiers).to.have.length(1);
-      expect(modifiers.indexOf('foo')).to.equal(0);
-    });
-    it('ignores falsy values', () => {
-      const modifiers = Helpers.getModifiersForDay(new Date(2015, 8, 19), {
-        foo: null,
-        bar: false,
-      });
-      expect(modifiers).to.have.length(0);
-      expect(modifiers.indexOf('foo')).to.equal(-1);
-    });
-    it('returns the modifier for an array of days', () => {
-      const modifiersObj = {
-        foo: [
-          new Date(2015, 8, 19),
-          new Date(2015, 8, 20),
-          new Date(2015, 8, 21),
-        ],
-        bar: [new Date(2015, 8, 19), new Date(2015, 8, 20)],
-      };
-      const modifiers1 = Helpers.getModifiersForDay(
-        new Date(2015, 8, 19),
-        modifiersObj
-      );
-      expect(modifiers1).to.have.length(2);
-      expect(modifiers1.indexOf('foo')).to.be.above(-1);
-      expect(modifiers1.indexOf('bar')).to.be.above(-1);
-
-      const modifiers2 = Helpers.getModifiersForDay(
-        new Date(2015, 8, 20),
-        modifiersObj
-      );
-      expect(modifiers2).to.have.length(2);
-      expect(modifiers2.indexOf('foo')).to.be.above(-1);
-      expect(modifiers2.indexOf('bar')).to.be.above(-1);
-
-      const modifiers3 = Helpers.getModifiersForDay(
-        new Date(2015, 8, 21),
-        modifiersObj
-      );
-      expect(modifiers3).to.have.length(1);
-      expect(modifiers3.indexOf('foo')).to.equal(0);
-      expect(modifiers3.indexOf('bar')).to.equal(-1);
-    });
-    it('accepts an array of days ignoring falsy values', () => {
-      const values = {
-        foo: [null, 'test', new Date(2015, 8, 21)],
-      };
-      const modifiers = Helpers.getModifiersForDay(
-        new Date(2015, 8, 21),
-        values
-      );
-      expect(modifiers).to.have.length(1);
-      expect(modifiers.indexOf('foo')).to.be.above(-1);
-    });
-    it('returns the modifier for a range of days', () => {
-      const range = {
-        foo: {
-          from: new Date(2015, 8, 18),
-          to: new Date(2015, 8, 20),
-        },
-      };
-      const modifiers1 = Helpers.getModifiersForDay(
-        new Date(2015, 8, 19),
-        range
-      );
-      expect(modifiers1).to.have.length(1);
-      expect(modifiers1.indexOf('foo')).to.equal(0);
-      const modifiers2 = Helpers.getModifiersForDay(
-        new Date(2015, 8, 17),
-        range
-      );
-      expect(modifiers2).to.have.length(0);
-    });
-    it('returns the modifier for multiple ranges of days', () => {
-      const ranges = {
-        foo: [
-          {
-            from: new Date(2015, 8, 18),
-            to: new Date(2015, 8, 20),
-          },
-          {
-            from: new Date(2015, 9, 18),
-            to: new Date(2015, 9, 20),
-          },
-        ],
-      };
-      const modifiers1 = Helpers.getModifiersForDay(
-        new Date(2015, 8, 19),
-        ranges
-      );
-      expect(modifiers1.indexOf('foo')).to.equal(0);
-      const modifiers2 = Helpers.getModifiersForDay(
-        new Date(2015, 9, 18),
-        ranges
-      );
-      expect(modifiers2.indexOf('foo')).to.equal(0);
-    });
-    it('returns an "after" modifier', () => {
-      const afterModifier = {
-        foo: {
-          after: new Date(2015, 8, 18),
-        },
-      };
-      const modifiers = Helpers.getModifiersForDay(
-        new Date(2015, 8, 19),
-        afterModifier
-      );
-      expect(modifiers).to.have.length(1);
-      expect(modifiers.indexOf('foo')).to.equal(0);
-    });
-    it('returns an "after" modifier in an array of days', () => {
-      const afterModifier = {
-        foo: [{ after: new Date(2015, 8, 18) }],
-      };
-      const modifiers = Helpers.getModifiersForDay(
-        new Date(2015, 8, 19),
-        afterModifier
-      );
-      expect(modifiers).to.have.length(1);
-      expect(modifiers.indexOf('foo')).to.equal(0);
-    });
-    it('returns a "before" modifier', () => {
-      const beforeModifier = {
-        foo: {
-          before: new Date(2015, 8, 15),
-        },
-      };
-      const modifiers = Helpers.getModifiersForDay(
-        new Date(2015, 8, 10),
-        beforeModifier
-      );
-      expect(modifiers).to.have.length(1);
-      expect(modifiers.indexOf('foo')).to.equal(0);
-    });
-    it('returns a "before" modifier in an array of days', () => {
-      const afterModifier = {
-        foo: [{ before: new Date(2015, 8, 15) }],
-      };
-      const modifiers = Helpers.getModifiersForDay(
-        new Date(2015, 8, 10),
-        afterModifier
-      );
-      expect(modifiers).to.have.length(1);
-      expect(modifiers.indexOf('foo')).to.equal(0);
-    });
-    it('works with mixing functions and other objects', () => {
-      const mixedModifiers = {
-        foo: [
-          { before: new Date(2015, 8, 15) },
-          day => day.getTime() === new Date(2015, 8, 17).getTime(),
-        ],
-      };
-      const modifiers = Helpers.getModifiersForDay(
-        new Date(2015, 8, 10),
-        mixedModifiers
-      );
-      expect(modifiers).to.have.length(1);
-      expect(modifiers.indexOf('foo')).to.equal(0);
-
-      const modifiers2 = Helpers.getModifiersForDay(
-        new Date(2015, 8, 17),
-        mixedModifiers
-      );
-      expect(modifiers2).to.have.length(1);
-      expect(modifiers2.indexOf('foo')).to.equal(0);
-    });
-    it('works even without modifiers', () => {
-      const modifiers = Helpers.getModifiersForDay(new Date(2015, 8, 19));
-      expect(modifiers).to.have.length(0);
-    });
-  });
-
   describe('getMonthsDiff', () => {
     it('returns a positive difference between two days in the same year', () => {
       const d1 = new Date(2015, 10, 6);
